feat(web): restructure marketplace integration detail page
- Extract the "What you can do" value-prop into a standalone
  IntegrationUseCases component and render it above the technical tool
  list (previously buried inside IntegrationRichContent).
- Add a collapsible ToolCard component with a 2-line description clamp and
  a "View more/View less" toggle that only renders when content overflows.
- Move the category/use-case fallback maps out of IntegrationRichContent
  into IntegrationUseCases (single owner of use-case generation).
- Replace em-dashes with commas across the SEO copy (descriptions, how-it-
  works steps, FAQs) for cleaner reading.
