+++ apps/web/src/app/[locale]/(landing)/marketplace/[slug]/page.tsx
@@ -156,19 +156,19 @@ function generateIntegrationDescription(integration: {
   const toolsNote = toolCount > 0 ? ` ${toolCount} actions available.` : "";
 
   const categoryDescriptions: Record<string, string> = {
-    email: `GAIA reads and triages your ${name} inbox, drafts context-aware replies, auto-labels threads, and creates tasks from emails — automatically, without manual prompts.${toolsNote}`,
-    calendar: `GAIA schedules meetings, prepares pre-meeting briefings, finds free slots, and manages your ${name} calendar proactively — triggered by email, conversation, or automation.${toolsNote}`,
-    task: `GAIA creates, updates, and prioritizes ${name} tasks automatically from your emails, calendar events, and conversations — no copy-pasting required.${toolsNote}`,
-    tasks: `GAIA creates, updates, and prioritizes ${name} tasks automatically from your emails, calendar events, and conversations — no copy-pasting required.${toolsNote}`,
-    productivity: `GAIA connects ${name} to your email, calendar, and workflows — automating the repetitive parts so you can focus on deep work.${toolsNote}`,
-    communication: `GAIA monitors your ${name} channels, summarizes threads, surfaces what needs your attention, and creates tasks from messages — automatically.${toolsNote}`,
-    crm: `GAIA updates ${name} contacts and deals automatically from email threads, meeting notes, and calendar events — keeping your CRM current without manual entry.${toolsNote}`,
-    notes: `GAIA saves meeting notes, email summaries, and task context directly to ${name} — automatically building your knowledge base without copy-pasting.${toolsNote}`,
+    email: `GAIA reads and triages your ${name} inbox, drafts context-aware replies, auto-labels threads, and creates tasks from emails, automatically, without manual prompts.${toolsNote}`,
+    calendar: `GAIA schedules meetings, prepares pre-meeting briefings, finds free slots, and manages your ${name} calendar proactively, triggered by email, conversation, or automation.${toolsNote}`,
+    task: `GAIA creates, updates, and prioritizes ${name} tasks automatically from your emails, calendar events, and conversations, no copy-pasting required.${toolsNote}`,
+    tasks: `GAIA creates, updates, and prioritizes ${name} tasks automatically from your emails, calendar events, and conversations, no copy-pasting required.${toolsNote}`,
+    productivity: `GAIA connects ${name} to your email, calendar, and workflows, automating the repetitive parts so you can focus on deep work.${toolsNote}`,
+    communication: `GAIA monitors your ${name} channels, summarizes threads, surfaces what needs your attention, and creates tasks from messages, automatically.${toolsNote}`,
+    crm: `GAIA updates ${name} contacts and deals automatically from email threads, meeting notes, and calendar events, keeping your CRM current without manual entry.${toolsNote}`,
+    notes: `GAIA saves meeting notes, email summaries, and task context directly to ${name}, automatically building your knowledge base without copy-pasting.${toolsNote}`,
     development: `GAIA creates ${name} issues from email threads, tracks PR status, prepares standup summaries, and syncs your engineering workflow with your calendar.${toolsNote}`,
-    automation: `GAIA triggers ${name} workflows from natural language — describe what you want to automate, and GAIA handles the execution across your connected tools.${toolsNote}`,
-    storage: `GAIA organizes files and documents in ${name} automatically — saving meeting notes, email attachments, and project artifacts in the right place.${toolsNote}`,
+    automation: `GAIA triggers ${name} workflows from natural language, describe what you want to automate, and GAIA handles the execution across your connected tools.${toolsNote}`,
+    storage: `GAIA organizes files and documents in ${name} automatically, saving meeting notes, email attachments, and project artifacts in the right place.${toolsNote}`,
     finance: `GAIA tracks ${name} activity, summarizes transactions, creates follow-up tasks from financial events, and integrates your financial data into your workflow.${toolsNote}`,
-    marketing: `GAIA automates ${name} tasks from email campaigns, tracks results, and creates follow-up workflows — connecting your marketing tool to your full productivity stack.${toolsNote}`,
+    marketing: `GAIA automates ${name} tasks from email campaigns, tracks results, and creates follow-up workflows, connecting your marketing tool to your full productivity stack.${toolsNote}`,
   };
 
   const categoryKey = integration.category?.toLowerCase();
@@ -176,7 +176,7 @@ function generateIntegrationDescription(integration: {
     return categoryDescriptions[categoryKey];
   }
 
-  return `Connect ${name} to GAIA and automate your workflows with AI. GAIA proactively manages your email, calendar, tasks, and connected tools — including ${name}.${toolsNote} Free MCP integration.`;
+  return `Connect ${name} to GAIA and automate your workflows with AI. GAIA proactively manages your email, calendar, tasks, and connected tools, including ${name}.${toolsNote} Free MCP integration.`;
 }
 
 function generateIntegrationTitle(name: string, category: string): string {
@@ -316,7 +316,7 @@ export default async function IntegrationPage({ params }: Props) {
 
   const howToSchema = generateHowToSchema(
     `How to automate ${integration.name} with GAIA`,
-    `Connect ${integration.name} to GAIA and start automating your ${categoryLabel.toLowerCase()} workflows in plain English — no code required.`,
+    `Connect ${integration.name} to GAIA and start automating your ${categoryLabel.toLowerCase()} workflows in plain English, no code required.`,
     [
       {
         name: `Connect ${integration.name} to GAIA`,
